src/fonctions/worker/requirements.txt
azure-core==1.39.0
azure-cosmos==4.15.0
azure-functions==1.24.0
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
MarkupSafe==3.0.3
requests==2.33.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8